feat(param): Begin To Add Language Stubs
see: https://plugins.jetbrains.com/docs/intellij/stub-indexes.html#implementation
Do note that with the way paramfiles are as a language, pretty much everything in the psi will be a stub.

TODO: ParamC is NOWHERE near finished. ParamBinaryFile has yet to be used as we are waiting for the stubs to get finished.